Refactory Validação

+++ src/main/java/com/br/estudo/crudpessoa/controller/PessoaController.java
@@ -1,5 +1,7 @@
 package com.br.estudo.crudpessoa.controller;
 
+import com.br.estudo.crudpessoa.controller.dto.PessoaDTO;
+import com.br.estudo.crudpessoa.controller.form.PessoaForm;
 import com.br.estudo.crudpessoa.model.Pessoa;
 import com.br.estudo.crudpessoa.repository.PessoaRepository;
 import com.br.estudo.crudpessoa.service.PessoaService;
@@ -8,6 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
@@ -22,33 +25,49 @@ public class PessoaController {
     @Autowired
     private PessoaService pessoaService;
 
-
     @GetMapping
-    public List<Pessoa> buscarTodos() {
-        return pessoaRepository.findAll();
-    }
-
-    @PostMapping
-    public ResponseEntity<Pessoa> criar(@RequestBody Pessoa pessoa) {
-        Pessoa pessoaSalva = pessoaRepository.save(pessoa);
-        return ResponseEntity.status(HttpStatus.CREATED).body(pessoaSalva);
+    public List<PessoaDTO> buscarTodos() {
+        return PessoaDTO.converter(pessoaRepository.findAll());
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Pessoa> buscarPeloCodigo(@PathVariable Long id) {
+    public ResponseEntity<PessoaDTO> buscarPeloCodigo(@PathVariable Long id) {
         Optional<Pessoa> pessoa = pessoaRepository.findById(id);
-        return pessoa.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+        return pessoa.map(value -> ResponseEntity.ok(new PessoaDTO(value))).orElseGet(() -> ResponseEntity.notFound().build());
+    }
+
+    @PostMapping
+    @Transactional
+    public ResponseEntity<PessoaDTO> criar(@Valid @RequestBody PessoaForm form) {
+        Pessoa pessoa = form.converter(null);
+        pessoa = pessoaService.salvarPessoa(pessoa);
+        return ResponseEntity.status(HttpStatus.CREATED).body(new PessoaDTO(pessoa));
+    }
+
+    @PutMapping("/{id}")
+    @Transactional
+    public ResponseEntity<PessoaDTO> atualizar(@PathVariable Long id, @Valid @RequestBody PessoaForm form) {
+        Optional<Pessoa> pessoaOptional = pessoaRepository.findById(id);
+
+        if (pessoaOptional.isPresent()) {
+            Pessoa pessoa = form.converter(id);
+            pessoa = pessoaService.editarPessoa(pessoaOptional.get(), pessoa);
+            return ResponseEntity.ok(new PessoaDTO(pessoa));
+        }
+
+        return ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void remover(@PathVariable Long id) {
-        pessoaRepository.deleteById(id);
-    }
+    @Transactional
+    public ResponseEntity<?> remover(@PathVariable Long id) {
+        Optional<Pessoa> ooptionalPessoa = pessoaRepository.findById(id);
 
-    @PutMapping
-    public ResponseEntity<Pessoa> atualizar(@RequestBody Pessoa pessoa) {
-        Pessoa pessoaSalva = pessoaRepository.save(pessoa);
-        return ResponseEntity.status(HttpStatus.OK).body(pessoaSalva);
+        if (ooptionalPessoa.isPresent()) {
+            pessoaRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+
+        return ResponseEntity.notFound().build();
     }
 }
